PARTIALLY FINISHED and NOT YET FINISHED
LOGIN AND SIGN UP ARE NOW PARTIALLY FINISHED

Reason? used basic fields first to ease the connection process. Previous made of database are not being used yet since its too many to implement all of it kinda overwhelming. So back to the small and basic fields until we get the hang out of it how some of the back-end works.

check user_info_V2 for more details